test: move #1227 concurrent-reload regression test and fix false-pass gap
Moves the concurrent reload_app regression test into TestReloadAppLocking
in test_app_lifecycle_service_operations.py, next to the sibling lock test
it complements, out of test_app_lifecycle_service_coverage.py (scoped to a
different kind of branch-coverage gap-filling).

Also adds a yield point in the mocked start-phase callback so the test
actually detects a regression that narrows the lock scope back to only the
stop phase -- without it, the mocked start body never suspended, so the
event loop never got a chance to interleave a second reload_app call during
the first's start phase, and the test passed even against the reintroduced
bug.
